+++ ftplugin/blueprint.lua
@@ -1,0 +1,9 @@
+-- Buffer Options
+vim.bo.tabstop = 4
+vim.bo.shiftwidth = 4
+vim.bo.textwidth = 80
+vim.bo.expandtab = true
+vim.bo.commentstring = "// %s"
+
+-- Window Options
+vim.wo.conceallevel = 2