+++ ftplugin/markdown.lua
@@ -1,0 +1,7 @@
+-- Buffer Options
+vim.bo.commentstring = "<!-- %s -->"
+
+-- Window Options
+vim.wo.wrap = true
+vim.wo.linebreak = true
+vim.wo.spell = true