+++ ftplugin/python.lua
@@ -1,0 +1,6 @@
+-- Buffer Options
+vim.bo.tabstop = 4
+vim.bo.shiftwidth = 4
+vim.bo.expandtab = true
+vim.bo.textwidth = 88  -- PEP8
+vim.bo.commentstring = "# %s"