+++ lua/plugins/indent-blankline.lua
@@ -1,0 +1,47 @@
+local highlight = {
+  "RainbowDelimiterRed",
+  "RainbowDelimiterYellow",
+  "RainbowDelimiterBlue",
+  "RainbowDelimiterOrange",
+  "RainbowDelimiterGreen",
+  "RainbowDelimiterViolet",
+  "RainbowDelimiterCyan",
+}
+
+return {
+  "lukas-reineke/indent-blankline.nvim",
+  event = "User FilePost",
+  opts = {
+    indent = { char = "│", highlight = "IblChar",},
+    scope = { char = "│", highlight = highlight },
+  },
+  config = function(_, opts)
+    dofile(vim.g.base46_cache .. "blankline")
+
+    local colors = dofile(vim.g.base46_cache .. "colors")
+    local hooks = require "ibl.hooks"
+
+    hooks.register(hooks.type.HIGHLIGHT_SETUP, function()
+      vim.api.nvim_set_hl(0, "RainbowDelimiterRed", { fg = colors.red })
+      vim.api.nvim_set_hl(0, "RainbowDelimiterYellow", { fg = colors.yellow })
+      vim.api.nvim_set_hl(0, "RainbowDelimiterBlue", { fg = colors.blue })
+      vim.api.nvim_set_hl(0, "RainbowDelimiterOrange", { fg = colors.orange })
+      vim.api.nvim_set_hl(0, "RainbowDelimiterGreen", { fg = colors.green })
+      vim.api.nvim_set_hl(0, "RainbowDelimiterViolet", { fg = colors.purple })
+      vim.api.nvim_set_hl(0, "RainbowDelimiterCyan", { fg = colors.cyan })
+    end)
+
+    hooks.register(
+      hooks.type.WHITESPACE,
+      hooks.builtin.hide_first_space_indent_level
+    )
+    
+    hooks.register(
+      hooks.type.SCOPE_HIGHLIGHT,
+      hooks.builtin.scope_highlight_from_extmark
+    )
+
+    require("ibl").setup(opts)
+    dofile(vim.g.base46_cache .. "blankline")
+  end,
+}