+++ lua/plugins/nvim-treesitter.lua
@@ -1,0 +1,29 @@
+return {
+  "nvim-treesitter/nvim-treesitter",
+  opts = {
+    ensure_installed = {
+      "bash",
+      "blueprint",
+      "css",
+      "dockerfile",
+      "gitignore",
+      "html",
+      "ini",
+      "javascript",
+      "json",
+      "htmldjango",
+      "lua",
+      "nix",
+      "python",
+      "requirements",
+      "rust",
+      "scss",
+      "sql",
+      "toml",
+      "vala",
+      "vim",
+      "vimdoc",
+      "yaml"
+    },
+  },
+}